python/requirements.txt
fdb==2.0.2
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2
pyyaml>=6.0




